Add files via upload
🌸 Flower Classification Dataset Description 🌸

Welcome to an exciting journey into the realm of Flower Classification! In this Machine Learning project, we delve into the enchanting world of flowers, utilizing a diverse dataset to create a robust classification model. in this project I have used many algorithms like Logistic Regression, Decision tree, Random forest, Bagging and Ada boosting or so on..



🧠🧠the project is focus on:-

📌Preprocessing of Data

📌Outlier treatment

📌Exploratory data analysis

📌Applying algorithms

📌Evaluation



🧠 Key Findings:

Logistic Regression and AdaBoost: Identified as the most effective algorithms for flower classification.

Cross-Validation: Ensured model reliability and generalizability.

Overfitting Concerns: Mitigated through careful algorithm selection and validation techniques.



🙏 Acknowledgments:

Special gratitude to @Kumar Sundram, @Mithilesh singh sir, whose guidance played a pivotal role in the success of this project. Thanks to @Learnbay for providing valuable sessions that equipped me with the knowledge to navigate and complete this endeavor successfully.

Heartfelt thanks to CodSoft for the opportunity for an enriching internship experience.





 #datascience #dataanalyst #flowerclassification #machinelearning #learnbay #codsoft #internshipopportunity 🌐